refactor(prisma): move prisma client to helpers
