+++ packages/prisma/__tests__/stringssession.ts
@@ -2,7 +2,7 @@ import { test, expect, describe, beforeEach } from 'vitest';
 import { session } from 'grammy';
 import { PrismaAdapter } from '../src';
 import { createBot, createMessage } from '@grammyjs/storage-utils';
-import prisma from '../src/prismaClient';
+import prisma from './helpers/prisma';
 
 beforeEach(async () => {
   await prisma.session.deleteMany({});
